+++ traits.h
@@ -1,31 +1,20 @@
 #ifndef __TRAITS_H__
 #define __TRAITS_H__
 
-#include <functional>
-
-// Forward declaration
-template <typename Traits>
-class CBinaryTreeNode;
-
-// Traits para nodos de árbol binario
-template <typename T>
-struct BinaryTreeNodeTraits {
+template <typename T, typename _Func>
+struct ListTrait{
     using value_type = T;
-};
-
-// Traits para árbol binario completo
-template <typename T>
-struct BinaryTreeAscTraits {
-    using value_type = T;
-    using CompareFn = std::less<T>;
-    using Node = CBinaryTreeNode<BinaryTreeNodeTraits<T>>;
+    using Func       = _Func;
 };
 
 template <typename T>
-struct BinaryTreeDescTraits {
-    using value_type = T;
-    using CompareFn = std::greater<T>;
-    using Node = CBinaryTreeNode<BinaryTreeNodeTraits<T>>;
+struct AscendingTrait : 
+    public ListTrait<T, std::less<T> >{
+};
+
+template <typename T>
+struct DescendingTrait : 
+    public ListTrait<T, std::greater<T> >{
 };
 
 #endif // __TRAITS_H__
